+++ CHANGELOG.md
@@ -7,6 +7,13 @@ This page shows a detailed overview of the changes between versions without the 
 	## __WORK IN PROGRESS__
 -->
 
+## __WORK IN PROGRESS__
+- Adjustment: Sort connection quality entries in network graph details by quality
+- Fix: Fixed encoding of custom array-based data types
+- Fix: Triggers a full node-update when a (Bridged Node) Basic Information cluster attribute changes
+- Fix: Uses the same kind of signal icons for all signal levels
+- Fix/Enhancement: Update matter.js to 0.16.11-nightly
+
 ## 0.4.2 (2026-02-22)
 - Fix: Sanitize wifi/thread credentials in the log when setting them
 - Fix: Update matter.js to 0.16.10